app root code optimization
